Changes: Overall: Changed some colors so it's not that ugly. Some small things that i can't recall.
On chinese_steak.html:
Added some divs to subdivide the page so it's easier to make stuff. ("image_container", "ingredients_section", "steps_section")
Added a shared box to the image and the recipe description, looks allright (not really).
Changed the first subtitle to a new class called "description", this is because the "description" h2 doesn't look good with the same box as other subtitles.
Added a "ingredients_item" class to each ingredient so later i can make changes to them.
some more small things that i can't recall

On styles.css:
Made a chain between "subtitle" and "description" class, so that h2 shares some similarities like the font and font-size but not the box-related stuff.
Made a box to the image, i had to add divs on the html because, to my surprise, i couldn't make a box out of an <img> so i had to make a div to make the box.
a lot more small things that i can't recall